+++ js/constant.js
@@ -21,8 +21,14 @@ const MAX_COUNT_OF_DECIMAL_NUMBERS = 5;
 const MAX_FEATURES = 6;
 const MAX_PHOTOS = 3;
 
+const TOKYO_LATITUDE = 35.6895;
+const TOKYO_LONGITUDE = 139.69171;
+const MAIN_PIN = 52;
+const PIN = 40;
+const START_POINTS = `${TOKYO_LATITUDE}, ${TOKYO_LONGITUDE}`;
+
 export {
   MIN_ELEMENTS, MIN_POSITIVE_NUMBER, COUNT_OF_MOCKS, MIN_LOCATION_X, MIN_LOCATION_Y,
   MAX_LOCATION_X, MAX_LOCATION_Y, MAX_PRICE, MAX_COUNT_OF_AVATARS, MAX_COUNT_OF_DECIMAL_NUMBERS,
-  MAX_FEATURES, MAX_PHOTOS, MIN_PRICE
+  MAX_FEATURES, MAX_PHOTOS, MIN_PRICE, TOKYO_LATITUDE, TOKYO_LONGITUDE, MAIN_PIN, PIN, START_POINTS
 };
